requirements.txt
python-telegram-bot
yt-dlp
yt-dlp-ejs
